CM - Aplicar execução de mais de uma página Google em cada execução da sync GSA.

+++ sme.gclass.api.dados/Interfaces/GoogleClassroom/IRepositorioCursoGsa.cs
@@ -8,7 +8,7 @@ namespace SME.GoogleClassroom.Dados
     {
         Task<bool> ExistePorIdAsync(string cursoId);
 
-        Task<long> SalvarAsync(CursoGsa cursoComparativo);
+        Task<int> SalvarAsync(CursoGsa cursoComparativo);
 
         Task<PaginacaoResultadoDto<CursoGsaDto>> ObterCursosComparativosAsync(Paginacao paginacao, string secao, string nome, string descricao);
